requirements.txt
python-dotenv
tavily-python
openai
